% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaRecur.R
\name{arimaRecur}
\alias{arimaRecur}
\title{arimaRecur}
\usage{
arimaRecur(data, order = c(1, 0, 1), mean = FALSE, steps, start)
}
\arguments{
\item{data}{Data to build the model from}

\item{order}{Optional 3-tuple specifiying order for the arima model}

\item{mean}{Optional boolean indicating whether to include a mean for the arima model}

\item{steps}{Numeric, how many steps ahead to forecast}

\item{start}{Numeric, where to start forecasting from in the data}
}
\description{
arimaRecur
}
\examples{
 arimaRecur(data, c(2,0,0), steps = 12, start = 100)

}
