% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbox.R
\name{quickbox}
\alias{quickbox}
\title{quickbox}
\usage{
quickbox(data)
}
\arguments{
\item{data}{Structured dataframe}
}
\description{
Outputs boxplots for all numeric variables in a dataset, good for preliminary descriptive analysi
}
