% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickhist.R
\name{quickhist}
\alias{quickhist}
\title{quickhist}
\usage{
quickhist(data)
}
\arguments{
\item{data}{Structured dataframe}
}
\value{
No return value, prints histograms for each variable in the dataframe
}
\description{
quickhist
}
