% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscores.R
\name{tscores}
\alias{tscores}
\title{tscores}
\usage{
tscores(model)
}
\arguments{
\item{model}{Fitted model from an Arima() call}
}
\description{
Returns tscores for all coefficients of an Arima model object
}
