% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointTest.R
\name{cointTest}
\alias{cointTest}
\title{cointTest}
\usage{
cointTest(t1, t2, start, end, T = 100, prices = FALSE)
}
\arguments{
\item{t1}{String, ticker for stock 1}

\item{t2}{String, ticker for stock 2}

\item{T}{Optional numeric, identifies how many days backward to include in the rolling window}

\item{prices}{Optional boolean for outputting the prices to a graph}
}
\description{
cointTest
}
