% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapeIter.R
\name{mapeIter}
\alias{mapeIter}
\title{mapeIter}
\usage{
mapeIter(forecasts, data, last = 1)
}
\arguments{
\item{forecasts}{Vector or dataframe of forecasts}

\item{data}{Dataframe of data to compare forecasts to}

\item{last}{Optional numeric, index of last point in data to build model, denotes where to start comparisons}
}
\description{
mapeIter
}
